Extract testing UUID related functionality.
